requirements.txt
pandas~=2.2.2
numpy~=1.26.4
streamlit~=1.33.0
seaborn~=0.13.2
scikit-learn~=1.4.2
matplotlib~=3.8.4
kmodes~=0.12.2
altair~=5.3.0
mysql-connector-python~=8.4.0
scipy~=1.13.0